+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 - Added "year" support for Holidays, which states if a holiday should only occur within one year (defaults to null).
 - Added "backgroundColor" support for Holidays, which states the background color that should be used for the day.
 - Added "textColor" support for Holidays, which states the text color that should be used for the day.
+- Added month name selection support to the Years Drop-Down on the main display (shows which month is currently selected).
 - 
 
 <br>